+++ CHANGELOG.md
@@ -11,10 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - `TransactionRecord.evmAddress`
  - `PublicKeyECDSA.toEvmAddress()`
  - `AccountCreateTransaction.setEvmAddress()`
+ - `AccountId.fromEvmAddress()`
+ - `AccountId.fromString()` now supports EVM address
+ - `TransferTransaction.addHbarTransfer()` now supports EVM address
  - `AccountCreateEvmAddressExample`
  - `CreateAccountWithAliasExample`
  - `LazyCreateAccountTransactionExample`
  - `LazyCreateTransferTransactionExample`
+ - `TransferUsingEvmAddressExample`
+ - `AccountCreationWaysExample`
 
 ### Deprecated